pyproject.toml
[project]
name = "shimmy-cloud-agents"
version = "0.1.0"
description = ""
authors = [
    {name = "Sam Silverberg"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk (>=1.0.0)",
    "google-cloud-aiplatform (>=1.92.0)",
    "google-cloud-speech (>=2.32.0,<3.0.0)",
    "google-cloud-texttospeech (>=2.26.0,<3.0.0)",
    "google-cloud-pubsub (>=2.29.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "grpcio (>=1.71.0,<2.0.0)",
    "grpcio-tools (>=1.71.0,<2.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "protobuf (>=6.31.1,<7.0.0)",
    "pytz (>=2024.1,<2025.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

